+++ core/jvm/src/main/kotlin/com/github/jcornaz/miop/experimental/Builders.kt
@@ -3,6 +3,7 @@ package com.github.jcornaz.miop.experimental
 import kotlinx.coroutines.experimental.Unconfined
 import kotlinx.coroutines.experimental.channels.ReceiveChannel
 import kotlinx.coroutines.experimental.channels.produce
+import kotlin.coroutines.experimental.CoroutineContext
 
 private val EmptyReceiveChannel = produce<Nothing>(Unconfined) { }
 
@@ -17,3 +18,15 @@ public fun <E> emptyReceiveChannel(): ReceiveChannel<E> = EmptyReceiveChannel
 public fun <E> receiveChannelOf(vararg values: E): ReceiveChannel<E> = produce(Unconfined, values.size) {
     for (value in values) send(value)
 }
+
+/**
+ * Return a [ReceiveChannel] which emits all elements of this iterable (in the same order)
+ */
+fun <T> Iterable<T>.openSubscription(context: CoroutineContext = Unconfined, capacity: Int = 0): ReceiveChannel<T> =
+        asSequence().openSubscription(context, capacity)
+
+/**
+ * Return a [ReceiveChannel] which emits all elements of this sequence (in the same order)
+ */
+fun <T> Sequence<T>.openSubscription(context: CoroutineContext = Unconfined, capacity: Int = 0): ReceiveChannel<T> =
+        produce(context, capacity = capacity) { forEach { send(it) } }